Update Ruby client
