+++ cdap-stream-clients/ruby/.gitignore
@@ -1,0 +1,2 @@
+*.lock
+pkg/*
