+++ cdap-stream-clients/ruby/.rubocop.yml
@@ -1,0 +1,6 @@
+AllCops:
+  Exclude:
+    - vendor/bundle/**/*
+# Disable specific cops below
+AsciiComments:
+  Enabled: false
