+++ cdap-stream-clients/ruby/Gemfile
@@ -1,4 +1,4 @@
-#  Copyright 2014-2015 Cask Data, Inc.
+#  Copyright © 2014-2016 Cask Data, Inc.
 #
 #  Licensed under the Apache License, Version 2.0 (the "License"); you may not
 #  use this file except in compliance with the License. You may obtain a copy of
@@ -12,21 +12,21 @@
 #  License for the specific language governing permissions and limitations under
 #  the License.
 
-source "http://rubygems.org"
+source 'http://rubygems.org'
 
 gemspec
 
-gem "rake"
+gem 'rake'
 
 gem 'httparty'
 gem 'thread'
-gem 'pry'
 
 group :test do
+  gem 'rubocop'
   gem 'rspec'
-  gem "simplecov"
-  gem "simplecov-rcov"
-  gem "vcr"
-  gem "webmock"
+  gem 'simplecov'
+  gem 'simplecov-rcov'
+  gem 'vcr'
+  gem 'webmock'
   gem 'cdap-authentication-client'
 end
