+++ cdap-stream-clients/ruby/spec/spec_helper.rb
@@ -1,19 +1,20 @@
-#  Copyright 2014 Cask Data, Inc.
+# Copyright © 2014-2016 Cask Data, Inc.
 #
-#  Licensed under the Apache License, Version 2.0 (the "License"); you may not
-#  use this file except in compliance with the License. You may obtain a copy of
-#  the License at
+# Licensed under the Apache License, Version 2.0 (the "License"); you may not
+# use this file except in compliance with the License. You may obtain a copy of
+# the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+# http://www.apache.org/licenses/LICENSE-2.0
 #
-#  Unless required by applicable law or agreed to in writing, software
-#  distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-#  WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-#  License for the specific language governing permissions and limitations under
-#  the License.
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+# License for the specific language governing permissions and limitations under
+# the License.
 
+# encoding: utf-8
 
-$:.push File.expand_path("../../lib", __FILE__)
+$LOAD_PATH.push File.expand_path('../../lib', __FILE__)
 
 require 'simplecov'
 require 'simplecov-rcov'
@@ -27,7 +28,7 @@ VCR.configure do |c|
   c.allow_http_connections_when_no_cassette = true
 end
 
-require 'stream-client-ruby'
+require 'cdap-stream-client'
 require 'rspec'
 
 Dir.glob("#{File.dirname(__FILE__)}/support/*.rb").each { |f| require f }
@@ -44,8 +45,8 @@ RSpec.configure do |config|
   # Run specs in random order to surface order dependencies. If you find an
   # order dependency and want to debug it, you can fix the order by providing
   # the seed, which is printed after each run.
-  #     --seed 1234  
-  config.order = "random"
-  
+  #     --seed 1234
+  config.order = 'random'
+
   # config.color_enabled = true
 end
